+++ src/app/js/config.js
@@ -73,6 +73,12 @@ var appConfig = {
     // Use `$uri` to refer to the ontology URI in srcUrlTemplate.
     ontViewers: [
       {
+        name: 'LODE',
+        title: 'Live OWL Documentation Environment',
+        srcUrlTemplate: 'http://www.essepuntato.it/lode/owlapi/$uri',
+        moreInfoUrl: 'http://www.essepuntato.it/lode'
+      },
+      {
         name: 'WebVOWL',
         title: 'Web-based Visualization of Ontologies',
         srcUrlTemplate: 'http://vowl.visualdataweb.org/webvowl/index.html#iri=$uri',